refactor(github): extract shared release-API helpers into crate::github
Pure cleanup so the upcoming in-app update checker (#118) can reuse the
same GitHub Releases primitives that bot::install already implements.

Behavior is unchanged. Moved to the new top-level module:
- GITHUB_API and USER_AGENT (akagi/<version>) constants.
- Asset and ReleaseJson serde structs, with new optional size and digest
  fields so the updater can verify GitHub's SHA-256 asset digest. Both
  default to None so the bot-install consumer is unaffected.
- build_client() and fetch_latest_release() so callers no longer
  hand-roll the reqwest builder + accept-header + JSON-decode chain.
- extract_zip_safe(), still with the zip-slip defence and unix mode
  preservation; the cft Chrome-for-Testing extractor also rewires to
  it via crate::github (the only other consumer outside bot::install).

cargo check --tests is clean; cargo test --lib passes 437/437.
